fix(callblasting): seconds con ceil (int), credits sin round #8
- CalculateDurationCustom: vuelve a ceil().cast(Int32) para seconds
- CalculateCreditsCallBlasting: elimina round(3) — créditos en valor exacto float
- CBSummaryGroup.seconds y CBSummaryGeneral.total_seconds: vuelven a int
- Tests restaurados con valores enteros para seconds

+++ src/modules/process/domain/models/summary.py
@@ -38,14 +38,14 @@ class CampaignSummary(BaseModel):
 class CBSummaryGroup(BaseModel):
     operator: str
     total: int
-    seconds: float
+    seconds: int
     credits: float
     unit_value: float
 
 
 class CBSummaryGeneral(BaseModel):
     total_records: int
-    total_seconds: float
+    total_seconds: int
     total_credits: float
     total_excluded: int
